requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Image processing
Pillow==10.1.0

# QR Code generation
qrcode[pil]==7.4.2

# WebSocket support
websockets==12.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0

# Web scraping for Ultimate Guitar integration
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
